snmp-monitor/requirements.txt
pysnmp==4.4.12
pyasn1==0.5.1
python-dotenv==1.0.0
schedule==1.2.0
